requirements.txt
# Install the local package so `src` can be imported.
-e .

# Core CLI script
click
python-dotenv

# Notebook and experiment workflow
jupyter
ipykernel
numpy
pandas
scipy
scikit-learn
nltk
xgboost
mlflow
dagshub
boto3

# Project tooling currently present in the repo
Sphinx
flake8
